refactor(wow-test): split Stages.kt into separate files for better maintainability
- Move CommandIterator to new file: CommandIterator.kt- Create new file for ExpectedResult: ExpectedResult.kt
- Extract ExpectStage interface to new file: ExpectStage.kt
- Create new file for WhenStage interface: WhenStage.kt
- Remove original Stages.kt file
